requirements.txt
telethon
python-dotenv
pydantic
httpx
rapidfuzz
aiogram==3.*  # для Telegram-бота управления
requests
